+++ src/Notifications/DailyChest.h
@@ -6,5 +6,5 @@ class DailyChest : public CCNode, public GJRewardDelegate {
     public:
         void getRewards(unsigned int type);
         void rewardsStatusFinished(int p0) override;
-        void rewardsStatusFailed() override {};
+        void rewardsStatusFailed() override;
 };